fix(share): open the full share modal from the sidebar and conversations list
The "Share" menu item in the sidebar's per-chat dropdown and on the conversations
page was copying the raw `/chat/<id>` URL to the clipboard — a private route that
only the owner can open — instead of opening the share modal used by the top
nav. That meant recipients got a link that 404'd for them.

Extract the share modal out of MainContent into a reusable component
(src/components/ShareModal) and render it from all three share entry points so
every "Share" affordance creates a real, public `/share/<token>` link with the
same create/copy/update/rotate/unshare controls.

The per-message "Share" button in MessageList is the export-as-file menu
(PDF/DOCX/etc.); unrelated to conversation sharing, left as-is.
