+++ src/components/ConversationsView/ConversationsView.module.css
@@ -877,45 +877,6 @@
   margin: 3px 6px;
 }
 
-/* Share confirmation icon */
-.confirmIconShare {
-  width: 48px;
-  height: 48px;
-  border-radius: 14px;
-  background: rgba(59, 130, 246, 0.08);
-  color: #3b82f6;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 0 auto 16px;
-}
-
-.confirmIconShare svg {
-  width: 20px;
-  height: 20px;
-}
-
-.confirmShareBtn {
-  flex: 1;
-  padding: 10px 16px;
-  border-radius: 10px;
-  font-size: 13px;
-  font-weight: 500;
-  background-color: #3b82f6;
-  color: white;
-  border: none;
-  cursor: pointer;
-  transition: background-color 0.15s ease, transform 0.1s ease;
-}
-
-.confirmShareBtn:hover {
-  background-color: #2563eb;
-}
-
-.confirmShareBtn:active {
-  transform: scale(0.98);
-}
-
 /* Load more */
 .loadMoreWrapper {
   display: flex;
