+++ src/components/Sidebar/Sidebar.module.css
@@ -1166,23 +1166,6 @@
   height: 20px;
 }
 
-.confirmIconShare {
-  width: 48px;
-  height: 48px;
-  border-radius: 14px;
-  background: rgba(59, 130, 246, 0.08);
-  color: #3b82f6;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 0 auto 16px;
-}
-
-.confirmIconShare svg {
-  width: 20px;
-  height: 20px;
-}
-
 .confirmTitle {
   font-size: 16px;
   font-weight: 600;
@@ -1237,27 +1220,6 @@
 }
 
 .confirmDeleteBtn:active {
-  transform: scale(0.98);
-}
-
-.confirmShareBtn {
-  flex: 1;
-  padding: 10px 16px;
-  border-radius: 10px;
-  font-size: 13px;
-  font-weight: 500;
-  background-color: #3b82f6;
-  color: white;
-  border: none;
-  cursor: pointer;
-  transition: background-color 0.15s ease, transform 0.1s ease;
-}
-
-.confirmShareBtn:hover {
-  background-color: #2563eb;
-}
-
-.confirmShareBtn:active {
   transform: scale(0.98);
 }
